backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document processing
python-docx==0.8.11
PyMuPDF==1.23.5

# Environment and utilities
python-dotenv==1.0.0

# LangChain and AI
langchain>=0.2.0
langchain-google-genai>=1.0.8

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1 

# Authentication
passlib==1.7.4
python-jose[cryptography]==3.4.0

# Email
sib-api-v3-sdk>=7.6.0
jinja2>=3.1.0

sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
alembic==1.12.0

# Authentication
bcrypt==4.0.1
argon2-cffi==23.1.0

# Email validation
email-validator==2.1.0